feat: restructure delete generator
